Update
Crypto coin price add Binance API handler
