+++ src/main/resources/memo.txt
@@ -2,12 +2,6 @@ AT 教程
 	rabbit MQ
 	redis
 	
-crypto coin
-	CryptoCoinNoticeConstant.HIT_NO_DATA_COUNTING_KEY
-		迁移至常量服务器
-	crypto compare 更新 daily data 时 可能返回全0数据, 需要binance 协助
-
-
 农历提醒
 图片信息传送
